+++ CHANGELOG.md
@@ -32,6 +32,7 @@
     - Improved consistency of `step_size` property across GD, ISTA, FISTA and APGD algorithms (#2157)
     - In PDHG algorithm, we now have options to initialise the dual variable, as well as the primal variable (#2169)
     - Allow FluxNormaliser.preview_configuration() even if flux contains zeros (#2177)
+    - Check if kwargs are used in AcquisitionData and ImageData initialisation (#2178)
   - Testing:
     - Developers can now run the full CI matrix [via the web UI](https://github.com/TomographicImaging/CIL/actions/workflows/build.yml) (#2160)
     - Added tests for ProjectionOperator inputs that use `unittest-parametrize` module (#1990)